refactor: redux 폴더 분리 및 폴더명 변경
