+++ src/index.js
@@ -4,7 +4,7 @@ import ReactDOM from 'react-dom/client';
 import 'index.css';
 
 import { Provider } from 'react-redux';
-import { store } from 'store/store';
+import { store } from 'redux/store';
 
 import App from 'App';
