+++ src/redux/store.js
@@ -1,7 +1,8 @@
 import { applyMiddleware, createStore } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import ReduxThunk from 'redux-thunk';
-import productsReducer from 'store/reducer/productsReducer';
+
+import productsReducer from 'redux/products/productsReducer';
 
 const store = createStore(productsReducer, composeWithDevTools(applyMiddleware(ReduxThunk)));
